reorg panther_default
