--- global_helpers/panther_default.yml
@@ -1,4 +1,0 @@
-AnalysisType: global
-GlobalID: "panther_default"
-Filename: panther_default.py
-Description: The default global accessible via the Panther web UI.
