+++ packs/aws.yml
@@ -194,7 +194,7 @@ PackDefinition:
     - panther_config
     - panther_config_defaults
     - panther_config_overrides
-    - panther_default
+    - panther_aws_helpers
     - panther_event_type_helpers
     - panther_iocs
     - panther_lookuptable_helpers
