+++ packs/credential_security.yml
@@ -20,7 +20,7 @@ PackDefinition:
     - global_filter_github
     - panther_auth0_helpers
     - panther_base_helpers
-    - panther_default
+    - panther_aws_helpers
     - panther_event_type_helpers
     - panther_config
     - panther_config_defaults
