+++ packs/standard_ruleset.yml
@@ -28,7 +28,7 @@ PackDefinition:
     - Standard.NewUserAccountCreated
     # Global Helpers
     - panther_base_helpers
-    - panther_default
+    - panther_aws_helpers
     - panther_config
     - panther_config_defaults
     - panther_config_overrides
