+++ policies/aws_cloudtrail_policies/aws_cloudtrail_s3_bucket_public.py
@@ -1,5 +1,5 @@
+from panther_aws_helpers import aws_regions
 from panther_base_helpers import deep_get
-from panther_default import aws_regions
 from panther_oss_helpers import BadLookup, resource_lookup
 
 BAD_PERMISSIONS = {
