+++ rules/aws_cloudtrail_rules/aws_cloudtrail_stopped.py
@@ -1,5 +1,4 @@
-from panther_aws_helpers import aws_rule_context
-from panther_default import aws_cloudtrail_success, lookup_aws_account_name
+from panther_aws_helpers import aws_cloudtrail_success, aws_rule_context, lookup_aws_account_name
 
 # API calls that are indicative of CloudTrail changes
 CLOUDTRAIL_STOP_DELETE = {
