+++ rules/aws_cloudtrail_rules/aws_console_login_without_mfa.py
@@ -1,7 +1,6 @@
 import logging
 
-from panther_aws_helpers import aws_rule_context
-from panther_default import lookup_aws_account_name
+from panther_aws_helpers import aws_rule_context, lookup_aws_account_name
 from panther_detection_helpers.caching import check_account_age
 
 # Set to True for environments that permit direct role assumption via external IDP
