+++ rules/aws_cloudtrail_rules/aws_console_login_without_saml.py
@@ -1,5 +1,4 @@
-from panther_aws_helpers import aws_rule_context
-from panther_default import lookup_aws_account_name
+from panther_aws_helpers import aws_rule_context, lookup_aws_account_name
 
 
 def rule(event):
