+++ rules/aws_cloudtrail_rules/aws_ec2_manual_security_group_changes.py
@@ -1,6 +1,5 @@
-from panther_aws_helpers import aws_rule_context
+from panther_aws_helpers import aws_cloudtrail_success, aws_rule_context
 from panther_base_helpers import pattern_match_list
-from panther_default import aws_cloudtrail_success
 
 PROD_ACCOUNT_IDS = {"11111111111111", "112233445566"}
 SG_CHANGE_EVENTS = {
