+++ rules/aws_cloudtrail_rules/aws_ec2_vpc_modified.py
@@ -1,5 +1,4 @@
-from panther_aws_helpers import aws_rule_context
-from panther_default import aws_cloudtrail_success
+from panther_aws_helpers import aws_cloudtrail_success, aws_rule_context
 
 # API calls that are indicative of an EC2 VPC modification
 EC2_VPC_MODIFIED_EVENTS = {
