+++ rules/aws_cloudtrail_rules/aws_ec2_vulnerable_xz_image_launched.py
@@ -1,5 +1,4 @@
-from panther_aws_helpers import aws_rule_context
-from panther_default import aws_cloudtrail_success
+from panther_aws_helpers import aws_cloudtrail_success, aws_rule_context
 from panther_iocs import XZ_AMIS
 
 # AMIs published by Fedora between 2024-03-26 and 2024-04-02
