+++ rules/aws_cloudtrail_rules/aws_kms_cmk_loss.py
@@ -1,5 +1,4 @@
-from panther_aws_helpers import aws_rule_context
-from panther_default import aws_cloudtrail_success
+from panther_aws_helpers import aws_cloudtrail_success, aws_rule_context
 
 # API calls that are indicative of KMS CMK Deletion
 KMS_LOSS_EVENTS = {"DisableKey", "ScheduleKeyDeletion"}
