+++ rules/aws_cloudtrail_rules/aws_resource_made_public.py
@@ -1,8 +1,7 @@
 import json
 
-from panther_aws_helpers import aws_rule_context
+from panther_aws_helpers import aws_cloudtrail_success, aws_rule_context
 from panther_base_helpers import deep_get
-from panther_default import aws_cloudtrail_success
 from policyuniverse.policy import Policy
