+++ rules/aws_cloudtrail_rules/aws_root_activity.py
@@ -1,4 +1,4 @@
-from panther_default import aws_cloudtrail_success, lookup_aws_account_name
+from panther_aws_helpers import aws_cloudtrail_success, lookup_aws_account_name
 
 EVENT_ALLOW_LIST = {"CreateServiceLinkedRole"}
