+++ rules/aws_cloudtrail_rules/aws_snapshot_made_public.py
@@ -1,8 +1,7 @@
 from collections.abc import Mapping
 
-from panther_aws_helpers import aws_rule_context
+from panther_aws_helpers import aws_cloudtrail_success, aws_rule_context
 from panther_base_helpers import deep_get
-from panther_default import aws_cloudtrail_success
 
 IS_SINGLE_USER_SHARE = False  # Used to adjust severity
