+++ rules/standard_rules/brute_force_by_ip.py
@@ -1,7 +1,7 @@
 from json import loads
 
 import panther_event_type_helpers as event_type
-from panther_default import lookup_aws_account_name
+from panther_aws_helpers import lookup_aws_account_name
 from panther_ipinfo_helpers import PantherIPInfoException, geoinfo_from_ip
 from panther_oss_helpers import add_parse_delay
